+++ src/tres/__init__.py
@@ -1,4 +1,4 @@
-from .tres import run_tres, run_tres_developer
-from .tres import initialize_triple_class
-from .tres import make_particle_sets
-from .tres import corotating_spin_angular_frequency_binary
+from .run_tres import run_tres, run_tres_developer
+from .run_tres import initialize_triple_class
+from .run_tres import make_particle_sets
+from .run_tres import corotating_spin_angular_frequency_binary
